fix overflow warning

+++ src/platform/nf-token/nf-token.h
@@ -29,7 +29,7 @@ namespace Platform
         /// Metadata buffer
         std::vector<unsigned char> metadata;
 
-        static const uint64_t UNKNOWN_TOKEN_PROTOCOL = -1;
+        static const uint64_t UNKNOWN_TOKEN_PROTOCOL = static_cast<uint64_t >(-1);
 
     public:
         ADD_SERIALIZE_METHODS
